Migrate and remove CommonOperationsTest
